chore(playground): add example of segfault when address expired
This is another case of segfault when the virtual memory address already
freed or not exist after the function return. One way to make sure
the virtual memory address still exist after the function return, we
can turn the variable into static variable, which still persist after
the function return.