chore(playground): add example of string overflow with array of char
Seems like there's a difference when using array of char with and
without the specified size.

When we give the specified size on array of char, like `char a1[3]`, we
need to give zero terminator at the end of the string literal so that
we don't get the value from other string literal.

But if we don't give the specified size on the array of char, like `char
a2[]`, we don't need to give zero terminator as the compiler already add
the zero terminator automatically (?).

This commit contains the example of string overflow with array of char.