chore(playground): add array in function parameter example
Seems like when we are passing array as function parameter, the array is
converted into pointer and when we are using `sizeof`, the value printed
is actually the size of pointer instead of the actual size of the array.

To solve this, we can add another function parameter for the array size
to calculate the length of array inside another function.

Please keep in mind that the array converted into pointer is only valid
for 1-dimensional array. If it's 2-dimensional array or more, the array
of arrays converted to "pointer of arrays" and not "pointer to
pointers".