feat(mpc): use array instead of variadic function in mpca_lang()
The problem with variadic function is that by default there's no
indicator to the end of the arguments passing into the function, so we
need to add the number that specifies the total arguments being passed
into variadic function.

The problem with adding the total arguments manually is that user can
input mismatch total arguments with the actual arguments being passed
into variadic function.

So, rather than using variadic function, let's try using array so that
we have a way to calculate the total arguments being passed rather than
input the total arguments manually. Let's see how it goes.