feat(sclisp): add error type and handle error division by zero
This is the first attempt implementing error handling. To make things
easier, we use struct that contains the value type (number or error),
the number value, and the error type.

As the example of how error handling works, in this changes i provide
the error handling of division by zero. Basically if we do `/ 69 0`, we
will get an error `Error: division by zero`.