feat(sclisp): use double instead of long integer for calculation
This is to support the calculation with decimal value such as 69.420.
This changes also remove support for decimal character using , (comma)
to make it more consistent with the expected decimal value of strtod()
which convert the input string into decimal value.

For now, i still have no idea how to trigger the overflow case with
strtod(), so the error handling for ERANGE might not work as expected.