refactor(run): use similar gcc flags for main and main-debug
main and main-debug subcommand have similarity in gcc flags used and the
arguments such as mpc.c and sclisp.c, the only difference is that
main-debug has flag -ggdb. So let's put the similar gcc flags and
arguments into a variable and reuse it for both main and main-debug
subcommand with main-debug subcommand adding extra -ggdb flag.