chore(playground): add strlen() info in string example
strlen() calculates the length of the string but exclude the null or
zero terminator (\0), so the actual length of the string is most likely
the return value of strlen() + 1.

Also, using the example string with array size but without null or zero
terminator (\0) with strlen(), result in the length of all characters
until we found a null or zero terminator (\0). That is interesting.