chore(playground): try replace printf() with write()
This changes is the first attempt to understand how can we print
`Holy C!` without any library. But for now, we are still using some
sort of library, which is `unistd.h`, so this is not the end goal.

I am committing this to document the step to print `Holy C!` without
any library.

At the moment, i still have no idea what the reference is all about, so
i will learn about that later on.