chore(playground): use embedded assembly to replace write()
This is the first attempt to not use any library. So instead of using
standard library, we use embedded assembly. At the moment, still have no
idea how assembly syntax is, so i need to learn that later on.

The problem with this approach is that the binary size is too big, it's
because the compiler still including the standard library in the binary
when we are using `-static` flag (?), at least that's according to the
youtube tutorial. So we will tackle that after this.

There's a subtle difference between this code and the tutorial code.
Instead of using `const char holy[]`, i am using `const char *holy` and
that lead to the change using `mov %0, %%rsi` instead of `lea %0,
%%rsi`. Still not sure what the difference between `mov` and `lea`, but
it seems like `mov` is the instruction to move data between registers
and memory? We will need to find that out later on.