refactor(lval_eval_sexpr): move count conditional at the top
This is so that we can immediately return if the count is 0 or 1 and not
going to the loop first.

Still not sure why we need to go to the loop first and then check the
count, because i see no difference whether we put the count conditional
before looping or vice versa.

Also, this changes combine the for-loop body as it has the same looping
condition.

Let's see how it goes.