feat(sclisp): implement modulo and power operator
Previously when we implement symbolic expression, we haven't implement
modulo and power operator so that we can try out the new symbolic
expression implementation.

Now that we have try out the new symbolic expression and haven't found
any problem yet, let's try implement modulo and power operator so that
we have the same feature as before.