chore(playground): use sizeof(*p) in dangling pointer example
Seems like this is more convenient as we only need to change the type
variable and `sizeof` will automatically get the size of the new type.
So we don't need to manually adjust the `sizeof` operand.

Also add explanation about `sizeof` and add example of assigning value
to pointer from malloc().