refactor(sclisp): use sizeof(*value) in malloc()
Just learned about using *value or the variable pointer as an operand
for sizeof so that we don't need the sizeof operand and just need to
change the variable value.

This is useful to prevent human-error when change the variable value for
malloc(), because we might forgot to change the sizeof operand in
malloc() too.