chore(playground): add pointer in function example
This is another example about using pointer in a function, something
like the continuation of dangling pointer example as we briefly mention
about using pointer of pointer inside the function.

This example is to understand the difference between changing the value
that the pointer points to and change the pointer itself when we are
passing the pointer into the function.