chore(playground): add integer widthness overflow example
Widthness in here means the maximum value of the specified integer type
such as int, unsigned int, long int, unsigned long int, and so on.

This is an example when the type can not hold the correct result value
and make the stored result truncated to fit the maximum value of the
type. This process causing a discrepancy between the correct result and
the stored result which can cause some unexpected behavior (a.k.a bug).