chore(playground): print pointer address in function parameter
To check if the address of pointer function parameter is the same as
the variable address used as function argument. It turns out the
function parameter address is different from the original variable.

It seems like we are copying the address from the pointer variable into
pointer in function parameter (?). Still not sure, need more info.