refactor(sclisp): not use malloc() and strcpy() in lval_err()
Because we already using string literal in lval_err(), that means we
already know the value before run-time. So it seems like we don't need
to use malloc() and strcpy() to give the error message.

This also reduce the need for memory allocation handling such as free
malloc after use and check if the malloc succeed or not.