fix(sclisp): can not evaluate the symbolic expression
Previously when we use symbolic expression like (/ 42 2), it will print
out the expression itself rather than evaluate the expression.

The problem is because when we are using symbolic expression, the first
iteration of the count is 1, similar to when we only put the number into
the interpreter.

If we have the conditional of count == 1 first, we will miss the
symbolic expression in the for loop and just print out the symbolic
expression as is.

So let's try move the conditional of count == 1 after the evaluation and
see how it goes.