feat(run): use static library for mpc parser
Because we rarely change the mpc parser source code, i feel like we
don't really need to re-compile mpc.c when trying to re-compile sclisp.

We only need to re-compile mpc.c when there's any changes to the parser.
Other than that, we just need to compile it once and re-use the
generated static library.

Using the static library for mpc parser make compiling sclisp faster
because the mpc parser source code is quite big and took some time to
compile.

To update the mpc parser, we can use the command `./run mpc`.