refactor: move enum, struct, and eval function to utils.{h,c}
Recently learned about C library development, and i want to try out
using header and C file like the C library code.

With that in mind, move out all enum and struct to utils.h and all
function prototype and function definition (except main function) to
utils.c.

Also, currently we don't use static library for utils.c, so we adjust
the run script to also compile utils.c.

I noticed that compiling using this method is a little bit slower
compared to compiling in one big files. Most likely because we are
including `mpc.h` in `utils.h` and we are including `utils.h` in both
`utils.c` and `sclisp.c` so twice the effort for the compiler (?). Still
not sure, need more info.

Let's see how it goes.