feat(mpc): exit program if parser not found
When there's parser not found when using mpca_lang(), we can get into
the REPL but the REPL will throw error "parser undefined" when we use
the REPL. I feel like it's better to just exit if there's a parser
that's can't be mapped.

Still not sure if we exit the program, do we still need to free the
heap memory manually or not?

Let's see how it goes.