refactor(mpc): use array of pointers for mpc_cleanup()
Rather than manually changed the parameters of mpc_cleanup() every time
we add a new parser, i feel like using array of parsers is more
convenient because we can calculate the array length automatically and
passing the array of parsers to mpc_cleanup() directly.

With this changes, basically we just need to add new parser into the
array of parsers and mpc_cleanup() will automatically pick up the new
value.

The concern of doing things this way is integer overflow when calculate
the array length, so we need to be careful to not add parsers more than
the INT_MAX in our machine.

Currently still not sure how to handle integer overflow correctly, maybe
will try handling integer overflow next time. We'll see.