chore(playground): add example for compound statement in for-loop
This example is to try out if we can assign the address of local
iterator variable in for-loop into pointer outside compound statement,
and it looks like we can do that (?).

By the time of this changes exist, i still have not found a reference
about the lifetime of compound statement inside a function.