feat(sclisp): add basic arithmetic operator in builtin()
Previously we use builtin_op() directly but now we have some keywords
like `head`, `list`, etc which is not the basic arithmetic operator.
Because of that, we create a new function called builtin() and put
the handler for those new keywords. But before this changes, we have not
add the basic arithmetic operator handler yet, which is builtin_op().

With that in mind, in this changes, we add builtin_op() into builtin().