chore(playground): add stack overflow example
This example supposed to be used with debugger like `gdb` to see how
many iteration we need to exhaust the stack limit.

On my system, it need 9 iteration to exhaust the stack limit. And the
interesting part is that when we use `continue <number>` command in
`gdb` and we got an error before finishing those `<number>`, when use
`run` command, we will be using the remainder of those `<number>` (?).

For example, when we use `continue 10` and got segfault in the 9
iteration, we will have one more execution when we do `run` command
again.

Still not sure if that hypothesis correct or not, that's just from my
observation. Need more info.