feat(sclisp): prevent print empty bracket when pressing only enter
Previously when we pressing only enter, without number or expression in
the prompt, we will get empty bracket as an output. This changes prevent
printing out empty bracket when there's no input.

But we still got empty line because of `putchar('\n')` in
lval_println(). Still not sure how to handle that yet.