chore(playground): add ksplice pointer challenge
This challenge introducing an interesting aspect of pointer, and that is
`&x + 1` which is different from `x + 1`, assuming x is an array.

It turns out that even though `x` and `&x` print out the same value,
which is the memory address of index 0, like `&x[0]`, it actually have
different result when we use arithmetic operator like `x + 1` and
`&x + 1`. That's because we use pointer arithmetic for the operation.