refactor(utils): not hardcoded 31 for integer overflow checker
Previously we hardcoded the right operand for shift right bitwise
operator with value 31, this is assuming that we use integer that always
32 bit. The possible problem is that when we change the `count` type to
another type, like `long long` type instead of `int`.

With that in mind, let's try using `sizeof` to get the `count` size and
calculate from that. We also try using `CHAR_BIT` just in case some
system have more or less than 8 bits in 1 byte.