chore(playground): change ksplice pointer challenge from int to char
Because char is 1 byte on my system, i feel like it's easier to
understand if we use `char` type instead of `int` type for this example.

Also add an explanation what happen when we do `&x+1` or `&x+2`. It's
basically when we use `&x`, it means we use the entire array and not
only the first element of the array. And because of that, when we do
`&x+1`, we are actually adding one element after the end of x array (in
this case is after the 5 elements).

And because we are using pointer in expression `&x+1`, we are actually
doing pointer arithmetic. Which we will take a look later.