chore(playground): add example of dereferencing 2-dimensional array
This example trying to check what expression `*confusion[1]` do, and it
seems like that expression is the same as `*(confusion[1])`. Basically
we are dereferencing the `confusion` memory address by offset 1 (the
next memory address) and then dereferencing those memory address again
(?). Not sure if that is the correct explanation but the point is that
the square bracket operator is executed __first__ and then the asterisk
operator.

We can also write those expression like this `*(*(confusion + 1))`.