chore: add explanation not using malloc() and strcpy() in lval_sym()
The book "Build Your Own Lisp" using malloc() and strcpy() to allocate
the string to new memory, and then i tried out not using malloc() and
strcpy() and just directly assign the memory address of `sym` function
parameter into the `value` memory.

Back then i didn't really understand why it works, but after analyzing
this for a little bit, i have a hypothesis why assigning the memory
address of `sym` function parameter directly into the `value` memory
works.

We might need to change this because the symbol in the chapter 11 used
for variable definition. Let's see how it goes.