refactor(utils): print out the same expression for s-expr and q-expr
Seems like in general lisp implementation we print out the quote
expression (q-expr) using the bracket symbol, similar to how we use
symbolic expression (s-expr).

For example, when we use symbolic expression, we need to give the
expression with enclosing bracket like `(+ 69 42)`. So when we use the
quote expression like `{69 42}`, we should print out the expression like
`(69 42)`.