refactor: exit program if there's memory allocation error
Previously i try to print out the `malloc failed` error message within
the REPL, but it turns out that mechanism can have the problem when
free the memory with `lval_del()` on complex expression.

Also, i want to use this with another memory allocation function such as
`realloc()` which is used in `lval_pop()` and that can get tricky to
handle in complex expression.

Currently have no idea how to handle the memory allocation error and
clean the occupied memory manually. So let's just exit the program if
there's a memory allocation error and let the OS handle the freeing
memory _for now_, at least until i figured out how to handle that
manually.