chore(playground): add explanation about type `char (*)[4]`
Finally able to _kind of_ understand what's going on with
`char (*)[4]`, which is kind of interesting that the type
`pointer to array` is actually pointing to the whole array instead of
the element inside of the array.

This kind of make me more understand that array is a separate type than
pointer. When we got an array type, it means we have _a whole array_
instead of just one of the _array's element_.